eas build fixes (#3971)
* fix expo-image

* fix eas

* cleanup
