+++ packages/app-mobile/app.config.ts
@@ -9,14 +9,14 @@ type ExpoExtras = {
 };
 
 const localGraphQLApi = "http://localhost:8080/v2/graphql";
-// const remoteGraphQLApi = "https://backpack-api.xnfts.dev/v2/graphql";
+const remoteGraphQLApi = "https://backpack-api.xnfts.dev/v2/graphql";
+
+const getUrl = (hash: string = "df0987c") =>
+  `https://mobile-service-worker.xnfts.dev/background-scripts/${hash}/service-worker-loader.html`;
 
 export default ({ config }: ConfigContext): ExpoConfig & ExpoExtras => {
   const projectID = "55bf074d-0473-4e61-9d9d-ecf570704635";
   const packageName = "app.backpack.mobile";
-
-  const getUrl = (hash: string = "1540e2d") =>
-    `https://mobile-service-worker.xnfts.dev/background-scripts/${hash}/service-worker-loader.html`;
 
   const remoteWebViewUrl = getUrl();
 
@@ -86,7 +86,10 @@ export default ({ config }: ConfigContext): ExpoConfig & ExpoExtras => {
       favicon: "./assets/favicon.png",
     },
     extra: {
-      graphqlApiUrl: localGraphQLApi,
+      graphqlApiUrl:
+        process.env.APP_ENV === "production"
+          ? remoteGraphQLApi
+          : localGraphQLApi,
       localWebViewUrl: "http://localhost:9333",
       remoteWebViewUrl,
       eas: {
