+++ packages/app-mobile/package.json
@@ -57,9 +57,10 @@
     "expo-dev-client": "~2.2.1",
     "expo-device": "~5.2.1",
     "expo-font": "~11.1.1",
-    "expo-image": "~1.0.0",
+    "expo-image": "~1.2.3",
     "expo-linear-gradient": "~12.1.2",
     "expo-linking": "~4.0.1",
+    "expo-local-authentication": "~13.3.0",
     "expo-random": "~13.1.1",
     "expo-secure-store": "~12.1.1",
     "expo-splash-screen": "~0.18.2",
