+++ packages/app-mobile/src/components/BottomSheetHelpModal.tsx
@@ -1,9 +1,13 @@
-import type { BottomSheetBackdropProps } from "@gorhom/bottom-sheet";
-
-import { useCallback, useEffect, useMemo, useRef } from "react";
-import { FlatList, Pressable, StyleSheet } from "react-native";
+import {
+  Alert,
+  DevSettings,
+  FlatList,
+  Pressable,
+  StyleSheet,
+} from "react-native";
 
 import * as Linking from "expo-linking";
+import { deleteItemAsync } from "expo-secure-store";
 
 import {
   BACKPACK_LINK,
@@ -11,13 +15,13 @@ import {
   TWITTER_LINK,
 } from "@coral-xyz/common";
 import { MaterialIcons } from "@expo/vector-icons";
-import { BottomSheetBackdrop, BottomSheetModal } from "@gorhom/bottom-sheet";
 
 import { BetterBottomSheet } from "~components/BottomSheetModal";
 import { DiscordIcon, TwitterIcon } from "~components/Icon";
-import { Margin, RoundedContainerGroup } from "~components/index";
+import { RoundedContainerGroup } from "~components/index";
 import { useTheme } from "~hooks/useTheme";
 import {
+  IconPushDetail,
   IconLaunchDetail,
   SettingsRow,
 } from "~screens/Unlocked/Settings/components/SettingsRow";
@@ -48,19 +52,60 @@ const styles = StyleSheet.create({
   },
 });
 
+const maybeResetApp = () => {
+  Alert.alert(
+    "Are your sure?",
+    "This will wipe all data that's been stored in the app",
+    [
+      {
+        text: "Yes",
+        onPress: () => {
+          doReset(true);
+        },
+      },
+      {
+        text: "No",
+        onPress: () => {},
+      },
+    ]
+  );
+};
+
+const doReset = async (shouldReset: boolean) => {
+  if (!shouldReset) {
+    return;
+  }
+  // TODO: don't manually specify this list of keys
+  const stores = [
+    "keyring-store",
+    "keyname-store",
+    "wallet-data",
+    "nav-store7",
+  ];
+  for (const store of stores) {
+    try {
+      await deleteItemAsync(store);
+    } catch (err) {
+      console.error(err);
+      // ignore
+    }
+  }
+
+  DevSettings.reload();
+};
+
 export function BottomSheetHelpModal({
   isVisible,
   resetVisibility,
-  extraOptions = [],
+  showResetButton,
 }: {
   isVisible: boolean;
   resetVisibility: () => void;
-  extraOptions?: any[];
+  showResetButton?: boolean;
 }): JSX.Element {
   const theme = useTheme();
 
   const menuOptions = [
-    ...extraOptions,
     {
       icon: (
         <MaterialIcons
@@ -86,6 +131,23 @@ export function BottomSheetHelpModal({
       detailIcon: <IconLaunchDetail />,
     },
   ];
+
+  if (showResetButton) {
+    menuOptions.push({
+      icon: (
+        <MaterialIcons
+          name="people"
+          size={24}
+          color={theme.custom.colors.secondary}
+        />
+      ),
+      label: "Reset Backpack",
+      detailIcon: <IconPushDetail />,
+      onPress: () => {
+        maybeResetApp();
+      },
+    });
+  }
 
   return (
     <BetterBottomSheet isVisible={isVisible} resetVisibility={resetVisibility}>
